Move tags resolver to its own package
